+++ src/components/flat-table/flat-table-row/flat-table-row.spec.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useRef } from "react";
 import { mount, ReactWrapper } from "enzyme";
 import { act } from "react-dom/test-utils";
 
@@ -17,11 +17,6 @@ import FlatTableRowHeader from "../flat-table-row-header/flat-table-row-header.c
 import FlatTableHeader from "../flat-table-header/flat-table-header.component";
 import { FlatTableBodyDraggable } from "..";
 import { FlatTableThemeContext } from "../flat-table.component";
-import guid from "../../../__internal__/utils/helpers/guid";
-
-const mockedGuid = "guid-12345";
-jest.mock("../../../__internal__/utils/helpers/guid");
-(guid as jest.MockedFunction<typeof guid>).mockImplementation(() => mockedGuid);
 
 const events = {
   enter: {
@@ -584,9 +579,6 @@ describe("FlatTableRow", () => {
           </thead>
         </table>
       );
-      act(() =>
-        wrapper.find(FlatTableHeader).at(0)?.props()?.reportCellWidth?.(200, 0)
-      );
 
       assertStyleMatch(
         {
@@ -641,13 +633,6 @@ describe("FlatTableRow", () => {
               </FlatTableRow>
             </thead>
           </table>
-        );
-        act(() =>
-          wrapper
-            .find(FlatTableHeader)
-            .at(1)
-            ?.props()
-            ?.reportCellWidth?.(200, 0)
         );
 
         assertStyleMatch(
@@ -1125,6 +1110,169 @@ describe("FlatTableRow", () => {
       });
     });
 
+    describe("when passing sub rows as a component", () => {
+      const SubRowsComponent = () => (
+        <>
+          <FlatTableRow>
+            <FlatTableCell>sub1cell1</FlatTableCell>
+            <FlatTableCell>sub1cell2</FlatTableCell>
+          </FlatTableRow>
+          <FlatTableRow>
+            <FlatTableCell>sub2cell1</FlatTableCell>
+            <FlatTableCell>sub2cell2</FlatTableCell>
+          </FlatTableRow>
+        </>
+      );
+
+      it("should expand the sub rows when row is clicked", () => {
+        const wrapper = renderFlatTableRow({
+          expandable: true,
+          subRows: <SubRowsComponent />,
+        });
+
+        act(() => {
+          wrapper.find(StyledFlatTableRow).at(0).props().onClick();
+        });
+
+        wrapper.update();
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(3);
+
+        act(() => {
+          wrapper.find(StyledFlatTableRow).at(0).props().onClick();
+        });
+
+        wrapper.update();
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(1);
+      });
+
+      it("should expand the sub rows when first column clicked and expandable area is first column", () => {
+        const wrapper = renderFlatTableRow({
+          expandable: true,
+          subRows: <SubRowsComponent />,
+          expandableArea: "firstColumn",
+        });
+
+        act(() => {
+          wrapper
+            .find(StyledFlatTableRow)
+            .find(StyledFlatTableCell)
+            .at(0)
+            .props()
+            .onClick();
+        });
+
+        wrapper.update();
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(3);
+
+        act(() => {
+          wrapper
+            .find(StyledFlatTableRow)
+            .find(StyledFlatTableCell)
+            .at(0)
+            .props()
+            .onClick();
+        });
+
+        wrapper.update();
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(1);
+      });
+
+      it("should toggle the open/close state of the row when enter/space pressed", () => {
+        const element = document.createElement("div");
+        const htmlElement = document.body.appendChild(element);
+
+        const wrapper = mount(
+          <table>
+            <tbody>
+              <FlatTableRow expandable subRows={<SubRowsComponent />}>
+                <FlatTableCell>cell1</FlatTableCell>
+                <FlatTableCell>cell2</FlatTableCell>
+              </FlatTableRow>
+            </tbody>
+          </table>,
+          { attachTo: htmlElement }
+        );
+
+        (wrapper
+          .find(StyledFlatTableRow)
+          .at(0)
+          .getDOMNode() as HTMLElement).focus();
+
+        act(() => {
+          wrapper
+            .find(StyledFlatTableRow)
+            .at(0)
+            .props()
+            .onKeyDown(events.enter);
+        });
+
+        wrapper.update();
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(3);
+
+        act(() => {
+          wrapper
+            .find(StyledFlatTableRow)
+            .at(0)
+            .props()
+            .onKeyDown(events.space);
+        });
+
+        wrapper.update();
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(1);
+      });
+
+      it("should toggle the open/close state of the row when enter/space pressed and expandable area is first column", () => {
+        const wrapper = renderFlatTableRow({
+          expandable: true,
+          subRows: <SubRowsComponent />,
+          expandableArea: "firstColumn",
+        });
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(1);
+
+        (wrapper
+          .find(StyledFlatTableRow)
+          .at(0)
+          .find("td")
+          .at(0)
+          .getDOMNode() as HTMLElement).focus();
+
+        act(() => {
+          wrapper
+            .find(StyledFlatTableRow)
+            .at(0)
+            .find(StyledFlatTableCell)
+            .at(0)
+            .props()
+            .onKeyDown(events.enter);
+        });
+
+        wrapper.update();
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(3);
+
+        act(() => {
+          wrapper
+            .find(StyledFlatTableRow)
+            .at(0)
+            .find(StyledFlatTableCell)
+            .at(0)
+            .props()
+            .onKeyDown(events.space);
+        });
+
+        wrapper.update();
+
+        expect(wrapper.find(StyledFlatTableRow).length).toEqual(1);
+      });
+    });
+
     describe.each([
       ["medium", "2px solid var(--colorsUtilityMajor100)"],
       ["large", "4px solid var(--colorsUtilityMajor100)"],
@@ -1302,9 +1450,6 @@ describe("FlatTableRow", () => {
             </FlatTableRow>
           </FlatTableBodyDraggable>
         </table>
-      );
-      act(() =>
-        wrapper.find(FlatTableHeader).at(0)?.props()?.reportCellWidth?.(200, 0)
       );
 
       wrapper.update();
@@ -1517,6 +1662,62 @@ describe("FlatTableRow", () => {
         { modifier: `${StyledFlatTableCell}` }
       );
     });
+
+    describe("with a ref", () => {
+      it("the ref should be forwarded", () => {
+        let mockRef = { current: null };
+
+        const WrapperComponent = () => {
+          mockRef = useRef(null);
+
+          return (
+            <table>
+              <FlatTableBodyDraggable>
+                <FlatTableRow id={0} ref={mockRef}>
+                  <FlatTableHeader>test 1</FlatTableHeader>
+                  <FlatTableCell>test 2</FlatTableCell>
+                  <FlatTableCheckbox />
+                  <FlatTableRowHeader>test 3</FlatTableRowHeader>
+                  <FlatTableHeader>test 4</FlatTableHeader>
+                  <FlatTableCell>test 5</FlatTableCell>
+                </FlatTableRow>
+              </FlatTableBodyDraggable>
+            </table>
+          );
+        };
+
+        const wrapper = mount(<WrapperComponent />);
+
+        expect(mockRef.current).toBe(wrapper.find("tr").getDOMNode());
+      });
+
+      it("the input callback ref should be called with the DOM element", () => {
+        let mockRef;
+
+        const WrapperComponent = () => {
+          mockRef = jest.fn();
+
+          return (
+            <table>
+              <FlatTableBodyDraggable>
+                <FlatTableRow id={0} ref={mockRef}>
+                  <FlatTableHeader>test 1</FlatTableHeader>
+                  <FlatTableCell>test 2</FlatTableCell>
+                  <FlatTableCheckbox />
+                  <FlatTableRowHeader>test 3</FlatTableRowHeader>
+                  <FlatTableHeader>test 4</FlatTableHeader>
+                  <FlatTableCell>test 5</FlatTableCell>
+                </FlatTableRow>
+              </FlatTableBodyDraggable>
+            </table>
+          );
+        };
+
+        const wrapper = mount(<WrapperComponent />);
+
+        expect(mockRef).toHaveBeenCalledWith(wrapper.find("tr").getDOMNode());
+      });
+    });
   });
 
   describe("wrapping FlatTableRowHeader", () => {
@@ -1526,7 +1727,6 @@ describe("FlatTableRow", () => {
       }: {
         children: React.ReactNode;
       }) => <FlatTableRowHeader>{children}</FlatTableRowHeader>;
-      FlatTableRowHeaderWrapper.displayName = FlatTableRowHeader.displayName;
       const rowHeaderIndex = mount(
         <table>
           <thead>
@@ -1545,6 +1745,42 @@ describe("FlatTableRow", () => {
         .prop("lhsRowHeaderIndex");
 
       expect(rowHeaderIndex).toEqual(3);
+    });
+  });
+
+  describe("when only children passed are checkboxes", () => {
+    it("does not update first cell index", () => {
+      const wrapper = mount(
+        <table>
+          <tbody>
+            <FlatTableRow>
+              <FlatTableCheckbox />
+              <FlatTableCheckbox />
+            </FlatTableRow>
+          </tbody>
+        </table>
+      );
+
+      expect(wrapper.find(StyledFlatTableRow).prop("firstCellIndex")).toBe(0);
+    });
+  });
+
+  describe("when first cell has no id set", () => {
+    it("does not set a first cell index", () => {
+      const wrapper = mount(
+        <table>
+          <tbody>
+            <FlatTableRow>
+              <td>I have no ID</td>
+              <FlatTableRowHeader>I have an ID</FlatTableRowHeader>
+            </FlatTableRow>
+          </tbody>
+        </table>
+      );
+
+      expect(wrapper.find(StyledFlatTableRowHeader).prop("leftPositon")).toBe(
+        undefined
+      );
     });
   });
 });